Componente card; Ajustes telas pedidos
