+++ src/pages/detalhespedido.jsx
@@ -1,17 +1,27 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect } from 'react'
 import Navbar from "../components/Navbar"
 import HeaderOrder from "../components/HeaderOrder"
 import Footer from '../components/Footer'
 import DetalhesAlternado from '../components/DetalhesAlternado'
 
 const detalhespedido = () => {
-    const [tituloPagina, setTituloPagina] = useState("Pedidos")
+    const [tituloPagina, setTituloPagina] = useState()
+    const [estagio, setEstagio] = useState(3)
+
+    function alternaTitulo(estagio){
+        estagio == 3 ? setTituloPagina("Pagamento")
+        : setTituloPagina("Pedidos") 
+    }
+
+    useEffect(() => {
+        alternaTitulo(estagio)
+      }, []);
 
     return (
         <>
             <Navbar fixed={true} page="/" logado={true} usuario="Ítalo" imgPerfil="/PerfilUsuario.jpeg" />
             <HeaderOrder tituloPagina={tituloPagina}/>
-            <DetalhesAlternado estagio={3}/>
+            <DetalhesAlternado estagio={estagio} />
             <Footer/>
         </>
     )
